Bug fixed: integer offside. (#185)

+++ src/VoltageSensor.cpp
@@ -1,6 +1,6 @@
 #include "VoltageSensor.h"
 
-VoltageSensor::VoltageSensor(int r1, int r2, int *const pins) : _factor((float) (r1 + r2) / r2), Device<float>(pins) {}
+VoltageSensor::VoltageSensor(float r1, float r2, int *const pins) : _factor((r1 + r2) / r2), Device<float>(pins) {}
 
 void VoltageSensor::initialize() { pinMode(_pins[0], INPUT); }
